feat: create prevision routes
