+++ src/controllers/prevision-controllers/types.ts
@@ -1,0 +1,3 @@
+import { Measurement } from "_/models";
+
+export type FindMeasurementQuery  = Partial<Measurement>
